Moved test config to top level so database is accessible for map integration tests
